% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec.R
\name{rec}
\alias{rec}
\title{Initialise a data pre-processing recipe}
\usage{
rec(data, y)
}
\arguments{
\item{data}{Data frame.}

\item{y}{The response variable.}
}
\value{
A object of class 'rec'. The object is a list that contains a data
 frame containing the meta data of the transformation and the results of some
 statistical tests. A data frame containing the actual transformed data. And
 the response variable.

 The columns in the meta data include:
  * id: The sequence id of the transformation
  * transformation: The transformation type
  * original_vars: The original variable(s) involved in the transformation
  * new_var: The name of the new variable as a result of the transformation
  * parameter: Any parameter used in the transformation, e.g. rolling window,
   leading, lagging order
  * norm_test: p-value from the Shapiro-Wilk’s normality test
  * station_test: p-value from the Ljung-Box test for stationarity
  * cor_resp: correlation coefficient with the target value.
}
\description{
Initialise a data pre-processing recipe
}
