% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_diff.R
\name{step_diff}
\alias{step_diff}
\title{Calculate the difference within a variable}
\usage{
step_diff(rec, o_var, n_var)
}
\arguments{
\item{rec}{A rec object}

\item{o_var}{Old variable to transform}

\item{n_var}{New variable name}
}
\value{
A rec object
}
\description{
Calculate the difference within a variable
}
\details{
Difference for a variable is defined as the nth value - the (n-1)th value
}
