% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_roll.R
\name{step_roll}
\alias{step_roll}
\title{Calculate rolling standard deviation}
\usage{
step_roll(rec, o_var, n_var, type = c("mean", "sd"), roll_window)
}
\arguments{
\item{rec}{A rec object}

\item{o_var}{Old variable to transform}

\item{n_var}{New variable name}

\item{type}{Mean or standard deviation}

\item{roll_window}{The rolling window as an integer.}
}
\value{
A rec object
}
\description{
Calculate rolling standard deviation
}
