% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_shift.R
\name{step_shift}
\alias{step_shift}
\title{Calculate lag and lead of a variable}
\usage{
step_shift(rec, o_var, n_var, type = c("lag", "lead"), shift_window)
}
\arguments{
\item{rec}{A rec object}

\item{o_var}{Old variable to transform}

\item{n_var}{New variable name}

\item{type}{Lag of lead}

\item{shift_window}{The shifting window (integer)}
}
\value{
A rec object
}
\description{
Calculate lag and lead of a variable
}
