% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_two.R
\name{step_two}
\alias{step_two}
\title{Calculate the spread, ratio and product of two variable}
\usage{
step_two(rec, o_vars, n_var, type = c("spread", "ratio", "product"))
}
\arguments{
\item{rec}{A rec object}

\item{o_vars}{A character vector of length 2, specifying the two variables to
include in the calculation.}

\item{n_var}{The name of the new variable}

\item{type}{Which calculation to perform, "spread", "ratio" or "product"}
}
\value{
A rec object
}
\description{
Calculate the spread, ratio and product of two variable
}
